Better tests for Jackson deserialization
