+++ main/src/main/java/com/gypsyengineer/ql/fun/jackson/one/SaferPersonDeserialization.java
@@ -40,8 +40,6 @@ public class SaferPersonDeserialization {
                 .polymorphicTypeValidator(ptv)
                 .build();
 
-        mapper.enableDefaultTyping();
-
         return mapper.readValue(string, clazz);
     }
